Alterado os tipos de dados double para decimal no ACBrFramework.Net - Rafael Batiati Corrigido o disparo de eventos nos componentes BAL e LCB  - Rafael Batiati

+++ ACBrFramework.Net/ACBrFramework.Net/ACBrAAC/ACBrAACECF.cs
@@ -6,7 +6,7 @@ namespace ACBrFramework
 	{
 		#region Properties
 
-		public double ValorGT { get; set; }
+		public decimal ValorGT { get; set; }
 
 		public string NumeroSerie { get; set; }
