+++ ACBrFramework.Net/ACBrFramework.Net/ACBrPAF/Registros/ACBrPAFRegistroR5.cs
@@ -8,27 +8,27 @@ namespace ACBrFramework
 
 		public string DESC_ITEM { get; set; }
 
-		public double QTDE_ITEM { get; set; }
+		public decimal QTDE_ITEM { get; set; }
 
 		public string UN_MED { get; set; }
 
-		public double VL_UNIT { get; set; }
+		public decimal VL_UNIT { get; set; }
 
-		public double DESCTO_ITEM { get; set; }
+		public decimal DESCTO_ITEM { get; set; }
 
-		public double ACRES_ITEM { get; set; }
+		public decimal ACRES_ITEM { get; set; }
 
-		public double VL_TOT_ITEM { get; set; }
+		public decimal VL_TOT_ITEM { get; set; }
 
 		public string COD_TOT_PARC { get; set; }
 
 		public string IND_CANC { get; set; }
 
-		public double QTDE_CANC { get; set; }
+		public decimal QTDE_CANC { get; set; }
 
-		public double VL_CANC { get; set; }
+		public decimal VL_CANC { get; set; }
 
-		public double VL_CANC_ACRES { get; set; }
+		public decimal VL_CANC_ACRES { get; set; }
 
 		public string IAT { get; set; }
